Added Response explainer, moved explaining code to CommandGrapher
